requirements.txt
streamlit
langchain
langchain-openai
langchain-groq
langchain-google-genai
google-generativeai
openai
python-dotenv
os